# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cake Delight API
  description: |
    REST API for the Cake Delight cloud native microservices application.

    Every endpoint is served through the API Gateway, which routes to the
    Catalog, Rating, Order, Inventory, Notification and Analytics services.

    Customer scoped endpoints require the `X-Customer-Id` header.
  version: 1.0.0
  license:
    name: ISC
servers:
  - url: http://localhost:8080
    description: Local API Gateway
...
components:
  parameters:
    CustomerId:
      name: X-Customer-Id
      in: header
      required: true
      description: Identifies the current customer.
      schema:
        type: string
      example: customer-1001
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
          nullable: true
    Cake:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
          example: CHOCOLATE
        price:
          type: number
          format: double
          example: 34.50
        currency:
          type: string
          example: USD
        imageUrl:
          type: string
        stockQuantity:
          type: integer
        available:
          type: boolean
        averageRating:
          type: number
          nullable: true
          description: Merged in from the Rating Service; null when unrated or unavailable.
        ratingCount:
          type: integer
          nullable: true
    CakePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Cake"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
...
    BasketItem:
      type: object
      properties:
        cakeId:
          type: string
        cakeName:
          type: string
        imageUrl:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
        lineTotal:
          type: number
    Basket:
      type: object
      properties:
        basketId:
          type: string
        customerId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BasketItem"
        itemCount:
          type: integer
        subtotal:
          type: number
        deliveryFee:
          type: number
        tax:
          type: number
        total:
          type: number
        currency:
          type: string
...
paths:
...
  /api/v1/cakes:
    get:
      tags: [Cakes]
      summary: List cakes with optional filters
      parameters:
        - name: name
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: search
          in: query
          description: Matches name or description.
          schema: { type: string }
        - name: available
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 12 }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, price, category, createdAt]
            default: name
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: includeRatings
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: A page of cakes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CakePage"
        "400":
          description: Invalid filter or paging values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Cakes]
      summary: Create a cake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cake"
      responses:
        "201":
          description: Cake created
        "400":
          description: Validation failed
...
  /api/v1/basket:
    get:
      tags: [Basket]
      summary: Get the current basket with calculated totals
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: The basket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Basket"
        "400":
          description: Missing X-Customer-Id
    delete:
      tags: [Basket]
      summary: Empty the basket
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: The empty basket
...
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List the customer's orders
      parameters:
        - $ref: "#/components/parameters/CustomerId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, PREPARING, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
      responses:
        "200":
          description: A page of orders
...